refactor(image-upload): extract to productImages router + harden clearMedia + 3 tests
