+++ backend/src/index.js
@@ -25,6 +25,7 @@ import marketingSpendRoutes from './routes/marketingSpend.js';
 import stockLossRoutes     from './routes/stockLoss.js';
 import publicRoutes        from './routes/public.js';
 import productRoutes       from './routes/products.js';
+import productImagesRouter from './routes/productImages.js';
 import stockOrderRoutes    from './routes/stockOrders.js';
 import floristHoursRoutes from './routes/floristHours.js';
 import premadeBouquetRoutes from './routes/premadeBouquets.js';
@@ -175,6 +176,7 @@ app.use('/api/intake',          intakeRoutes);
 app.use('/api/settings',        settingsRoutes);
 app.use('/api/marketing-spend', marketingSpendRoutes);
 app.use('/api/stock-loss',      stockLossRoutes);
+app.use('/api/products',        productImagesRouter);
 app.use('/api/products',        productRoutes);
 app.use('/api/stock-orders',    stockOrderRoutes);
 app.use('/api/florist-hours',  floristHoursRoutes);
